+++ src/dtgtk/culling.c
@@ -1403,8 +1403,8 @@ static gboolean _thumbs_compute_positions(dt_culling_t *table)
   {
     dt_thumbnail_t *thumb = l->data;
     const float aspect_ratio = thumb->aspect_ratio;
-    thumb->width = (gint)(sqrt(aspect_ratio) * 100);
-    thumb->height = (gint)(1 / sqrt(aspect_ratio) * 100);
+    thumb->width = (gint)(sqrtf(aspect_ratio) * 100);
+    thumb->height = (gint)(1 / sqrtf(aspect_ratio) * 100);
     thumb->x = thumb->y = 0;
 
     max_thumb_height = MAX(max_thumb_height, thumb->height);
